MTA Navigator/requirements.txt
streamlit>=1.28.0
fastapi==0.115.5
uvicorn==0.32.1
google-generativeai>=0.8.5
networkx==3.2.1
pandas==2.2.3
sqlalchemy==2.0.23
rapidfuzz==3.6.1
requests==2.32.5
python-multipart==0.0.6
python-dotenv==1.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.27.0
gtfs-realtime-bindings>=2.0.0
